+++ executor-service/src/main/java/executor/service/maintenance/ScenarioExecutor.java
@@ -1,4 +1,4 @@
-package executor.service.maintenance.plugin.proxy;
+package executor.service.maintenance;
 
 import executor.service.model.ScenarioDto;
 import org.openqa.selenium.WebDriver;
