+++ executor-service/src/test/java/executor/service/ParallelFlowExecutorImplTest.java
@@ -1,8 +1,8 @@
 package executor.service;
 
 import executor.service.factory.webdriverinitializer.WebDriverProvider;
+import executor.service.maintenance.ScenarioExecutor;
 import executor.service.maintenance.plugin.proxy.ProxySourcesClient;
-import executor.service.maintenance.plugin.proxy.ScenarioExecutor;
 import executor.service.model.ProxyConfigHolderDto;
 import executor.service.model.ThreadPoolConfigDto;
 import org.junit.jupiter.api.BeforeEach;
